packages/corp-opportunity-manager/pyproject.toml
[project]
name = "corp-opportunity-manager"
version = "0.2.0"
description = "Opportunity lifecycle management for pre-sales workflows"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "openpyxl>=3.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
graph = ["msal>=1.24", "httpx>=0.25"]
llm = ["google-genai>=1.0"]
dev = ["pytest>=7.0", "pytest-cov", "ruff"]

[project.scripts]
com = "corp_opportunity_manager.cli:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
